Redone the whole parser, made it simpler
I redid the logic quite a bit to make it a bit simpler to understand.
The `process` function still has a lot of content inside it which might
be nicer to split up. Otherwise, it is debatable whether to have:

Query {name, length, hits (Vec<SigAl>)}

OR

SigAl {info, length, score, evalue, origin(Query)}

I think the second one is better if we are working purely with the SigAl
i.e sorting them and so on.

If we wanted to go by query, and then have the SigAl as a vec in the
Query object i.e so each query has a list of alignments which are
significant, we can easily do that as well.